Squashed commit of the following:
commit b93818d0e57aff11938d7ad9ebbf247967a007c5
Author: Rawaa Alatrash <rawaa.alatrash@gmail.com>
Date:   Fri Mar 3 20:55:16 2023 +0100

    material-video

+++ webapp/src/app/pages/components/materils/material/material.component.ts
@@ -101,11 +101,13 @@ export class MaterialComponent implements OnInit{
   onTabChange(e) {
     this.tabIndex = e.index - 1;
     if (this.tabIndex == -1) this.isMaterialSelected = false;
+   // if (this.tabIndex == 0) this.isMaterialSelected = false;
     else this.isMaterialSelected = true;
 
-    this.setSelectedTabIndex(this.tabIndex || 0);
+    this.setSelectedTabIndex(this.tabIndex );
 
-    this.selectedMaterial = this.channels['materials'][this.tabIndex || 0];
+    this.selectedMaterial = this.channels['materials'][this.tabIndex ];
+    
 
     this.router.navigate([
       'course',
@@ -118,6 +120,9 @@ export class MaterialComponent implements OnInit{
     this.store.dispatch(
       MaterialActions.setMaterialId({ materialId: this.selectedMaterial._id })
     );
+    this.store.dispatch(
+      MaterialActions.setCurrentMaterial({ selcetedMaterial: this.selectedMaterial })
+    );
     this.store.dispatch(AnnotationActions.loadAnnotations());
   }
   setSelectedTabIndex(index: number) {
@@ -126,12 +131,36 @@ export class MaterialComponent implements OnInit{
   }
   updateSelectedMaterial() {
     if (!this.selectedMaterial) return;
+    console.log("this.selectedMaterial  print whole material details")
+    console.log(this.selectedMaterial)
     switch (this.selectedMaterial.type) {
       case 'pdf':
         this.pdfViewService.setPageNumber(1);
         let url =
           this.selectedMaterial?.url + this.selectedMaterial?._id + '.pdf';
         this.pdfViewService.setPdfURL(url);
+        console.log("url is empty for pdf")
+
+        break;
+        case 'video':
+        if(this.selectedMaterial?.url == "")
+        {
+          
+          let urlvideo =
+          "/public/uploads/videos/" + this.selectedMaterial?._id + '.mp4';
+           this.pdfViewService.setPdfURL(urlvideo);
+           console.log(urlvideo)
+           console.log("url is empty")
+
+          
+        }
+        else{   
+
+          this.pdfViewService.setPdfURL(this.selectedMaterial?.url);
+          console.log(this.selectedMaterial?.url)
+          console.log("url is not empty")
+      }
+
 
         break;
 
@@ -146,16 +175,68 @@ export class MaterialComponent implements OnInit{
     e.index1 = e.index - 1;
 
     this.selectedMaterial = this.channels['materials'][e.index1];
+    console.log("this.selectedMaterial with index")
+    console.log(this.selectedMaterial)
+    if(this.selectedMaterial.type == "video" && this.selectedMaterial.url )
+    {
+      this.materialService.deleteMaterial(this.selectedMaterial).subscribe({
+        next: (data) => {
+          console.log(data);
+         
+  
+     
+          this.topicChannelService.selectChannel(this.selectedChannel)
+  this.router.navigate([
+    'course',
+    this.selectedMaterial['courseId'],
+    'channel',
+    this.selectedMaterial['channelId'],
+    
+  ]); 
+  e.index = 1
+  
+  
+   console.log("material deleted")
+  
+        },
+        error: (err) => {
+          this.errorMessage = err.error.message;
+        },
+      });
+    }
+
+    else  if((this.selectedMaterial.type == "pdf" && this.selectedMaterial.url )||(this.selectedMaterial.type == "video" && this.selectedMaterial.url== "" ) )
+    {   
+       this.materialService.deleteFile(this.selectedMaterial).subscribe({
+              next: (res) => {
+                console.log(res);
+                console.log("file material deleted")
+              },
+            });
+
+      }
+
+
 
     this.materialService.deleteMaterial(this.selectedMaterial).subscribe({
       next: (data) => {
         console.log(data);
+       
 
-        this.materialService.deleteFile(this.selectedMaterial).subscribe({
-          next: (res) => {
-            console.log(res);
-          },
-        });
+   
+        this.topicChannelService.selectChannel(this.selectedChannel)
+this.router.navigate([
+  'course',
+  this.selectedMaterial['courseId'],
+  'channel',
+  this.selectedMaterial['channelId'],
+  
+]); 
+
+e.index = 1
+
+ console.log("material deleted")
+
       },
       error: (err) => {
         this.errorMessage = err.error.message;
